% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{as_open}
\alias{as_open}
\title{Convert a catlab StructuredCospan back to an Open Petri net}
\usage{
as_open(sc)
}
\arguments{
\item{sc}{A catlab::StructuredCospan}
}
\value{
An Open Petri net
}
\description{
Extracts the apex and recovers each leg as an integer vector of species
indices from the ACSetTransformation components.
}
