% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{as_structured_cospan}
\alias{as_structured_cospan}
\title{Convert an Open Petri net to a catlab StructuredCospan}
\usage{
as_structured_cospan(open_pn)
}
\arguments{
\item{open_pn}{An Open Petri net}
}
\value{
A catlab::StructuredCospan with interface_ob = "S"
}
\description{
Each integer-vector leg becomes an ACSetTransformation from a discrete
Petri net foot (with parts only in "S") to the apex.
}
