% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{build_machine_closure}
\alias{build_machine_closure}
\title{Build a dynamics closure from symbolic expressions for a Machine}
\usage{
build_machine_closure(state_names, ninputs, dynamics_expr)
}
\arguments{
\item{state_names}{Character vector of state variable names}

\item{ninputs}{Number of input ports}

\item{dynamics_expr}{Named list of quoted expressions}
}
\value{
A function(u, x, p, t) -> du
}
\description{
Build a dynamics closure from symbolic expressions for a Machine
}
\keyword{internal}
