% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{build_prior}
\alias{build_prior}
\title{Build a prior model from specifications}
\usage{
build_prior(...)
}
\arguments{
\item{...}{Named prior specifications}
}
\value{
A monty_model object
}
\description{
Constructs a monty_model for use as a prior in \code{\link[=fit_mcmc]{fit_mcmc()}}.
Each argument is a named prior specification created by
\code{\link[=prior_exp]{prior_exp()}}, \code{\link[=prior_uniform]{prior_uniform()}}, or \code{\link[=prior_normal]{prior_normal()}}.
}
\examples{
\donttest{
p <- build_prior(
  beta = prior_exp(mean = 0.3),
  gamma = prior_exp(mean = 0.1)
)
}
}
