% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{build_rs_closure}
\alias{build_rs_closure}
\title{Build a dynamics closure from symbolic expressions}
\usage{
build_rs_closure(state_names, dynamics_expr, param_names)
}
\arguments{
\item{state_names}{Character vector of state variable names}

\item{dynamics_expr}{Named list of quoted expressions}

\item{param_names}{Character vector of parameter names}
}
\value{
A function(u, p, t) -> du suitable for ResourceSharer
}
\description{
Build a dynamics closure from symbolic expressions
}
\keyword{internal}
