% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{build_stratified_flow_expr}
\alias{build_stratified_flow_expr}
\title{Build stratified flow rate expression}
\usage{
build_stratified_flow_expr(
  sf1,
  sf2,
  fn1,
  fn2,
  r1,
  r2,
  sn1,
  sn2,
  pn1,
  pn2,
  sp_pair,
  independent_params = FALSE
)
}
\description{
Build stratified flow rate expression
}
\keyword{internal}
