% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{delay_machine}
\alias{delay_machine}
\title{Create a DelayMachine}
\usage{
delay_machine(
  ninputs,
  nstates,
  noutputs,
  dynamics,
  readout,
  state_names = paste0("x", seq_len(nstates)),
  dynamics_expr = list(),
  readout_expr = list(),
  params = character(0)
)
}
\arguments{
\item{ninputs}{Number of input ports}

\item{nstates}{Number of internal states}

\item{noutputs}{Number of output ports}

\item{dynamics}{Function (u, x, h, p, t) -> du}

\item{readout}{Function (u, h, p, t) -> y}

\item{state_names}{Optional state names}

\item{dynamics_expr}{Optional named list of expressions (for code gen)}

\item{readout_expr}{Optional list of readout expressions}

\item{params}{Parameter names}
}
\description{
Create a DelayMachine
}
