% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{dwd}
\alias{dwd}
\title{Create a directed wiring diagram for Machine composition}
\usage{
dwd(ninputs, noutputs, boxes, wires)
}
\arguments{
\item{ninputs}{Number of external input ports}

\item{noutputs}{Number of external output ports}

\item{boxes}{Named list where each element is c(ninputs, noutputs)}

\item{wires}{List of wire specs: c(source_box, source_port, target_box, target_port).
Use 0 for external input box, -1 for external output box.}
}
\value{
A DWD list structure
}
\description{
Create a directed wiring diagram for Machine composition
}
