% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{euler_approx}
\alias{euler_approx}
\title{Convert a continuous system to discrete via Euler's method}
\usage{
euler_approx(rs, h = NULL)
}
\arguments{
\item{rs}{A ResourceSharer or Machine with system_type = "continuous"}

\item{h}{Step size (numeric). If NULL, step size is appended to parameters.}
}
\value{
A discrete system of the same class
}
\description{
Convert a continuous system to discrete via Euler's method
}
