% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{eval_dynamics}
\alias{eval_dynamics}
\title{Evaluate dynamics of a ResourceSharer}
\usage{
eval_dynamics(rs, ...)
}
\arguments{
\item{rs}{A ResourceSharer}

\item{...}{Additional arguments passed to methods}
}
\description{
Evaluate dynamics of a ResourceSharer
}
