% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{eval_readout}
\alias{eval_readout}
\title{Evaluate readout of a Machine}
\usage{
eval_readout(m, u, p = NULL, t = 0, h = NULL)
}
\arguments{
\item{m}{A Machine}

\item{u}{State vector}

\item{p}{Parameter list}

\item{t}{Time}

\item{h}{History function (for delay systems)}
}
\value{
Readout vector
}
\description{
Evaluate readout of a Machine
}
