% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{exposed_states}
\alias{exposed_states}
\title{Get exposed state values at ports}
\usage{
exposed_states(rs, u)
}
\arguments{
\item{rs}{A ResourceSharer}

\item{u}{State vector}
}
\value{
Numeric vector of exposed state values
}
\description{
Get exposed state values at ports
}
