% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{flatten}
\alias{flatten}
\title{Extract the underlying Petri net from a TypedPetriNet}
\usage{
flatten(typed_pn)
}
\arguments{
\item{typed_pn}{A TypedPetriNet}
}
\value{
A Petri net ACSet
}
\description{
Extract the underlying Petri net from a TypedPetriNet
}
