% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{flow}
\alias{flow}
\title{Create a flow specification}
\usage{
flow(from = NULL, to = NULL, rate)
}
\arguments{
\item{from}{Source stock name (NULL or "cloud" for external inflow)}

\item{to}{Target stock name (NULL or "cloud" for external outflow)}

\item{rate}{Rate expression (quoted) or function(u, p, t)}
}
\value{
A flow spec list
}
\description{
Create a flow specification
}
