% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{infectious_ontology}
\alias{infectious_ontology}
\title{Standard infectious disease ontology}
\usage{
infectious_ontology()
}
\value{
A LabelledPetriNet ACSet
}
\description{
Returns a type system (LabelledPetriNet) with:
\itemize{
\item One species type: \code{Pop}
\item Transition types: \code{infect} (Pop,Pop -> Pop,Pop),
\code{disease} (Pop -> Pop), \code{strata} (Pop -> Pop)
}
}
\examples{
onto <- infectious_ontology()
species_names(onto)    # "Pop"
transition_names(onto) # c("infect", "disease", "strata")
}
