% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\name{LabelledPetriNet}
\alias{LabelledPetriNet}
\title{LabelledPetriNet ACSet type}
\usage{
LabelledPetriNet(...)
}
\arguments{
\item{...}{Arguments passed to the ACSet constructor}
}
\description{
LabelledPetriNet ACSet type
}
