% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{Machine}
\alias{Machine}
\title{A machine: open dynamical system with directed inputs and outputs.}
\usage{
Machine(
  system_type = "continuous",
  ninputs = 0L,
  nstates = 0L,
  noutputs = 0L,
  dynamics = function(u, x, p, t) numeric(0),
  readout = function(u, p, t) u,
  state_names = character(0),
  dynamics_expr = list(),
  readout_expr = list(),
  params = character(0)
)
}
\arguments{
\item{system_type}{One of "continuous", "discrete", "delay"}

\item{ninputs}{Number of input ports}

\item{nstates}{Number of internal state variables}

\item{noutputs}{Number of output ports}

\item{dynamics}{Function: \verb{(u, x, p, t) -> du} or \verb{(u, x, h, p, t) -> du} for delay}

\item{readout}{Function: \verb{(u, p, t) -> y} or \verb{(u, h, p, t) -> y} for delay}

\item{state_names}{Character vector of state variable names}

\item{dynamics_expr}{List of symbolic expressions for code generation}

\item{readout_expr}{List of symbolic readout expressions}

\item{params}{Character vector of parameter names}
}
\description{
Corresponds to AlgebraicDynamics.jl's Machine. Has explicit input ports
(receiving signals) and output ports (emitting readouts).
}
