% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{merge_machine_expressions}
\alias{merge_machine_expressions}
\title{Merge symbolic expressions during oapply_machine composition}
\usage{
merge_machine_expressions(d, machines, all_snames, offsets, state_counts)
}
\description{
Merge symbolic expressions during oapply_machine composition
}
\keyword{internal}
