% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{merge_machine_readout_expr}
\alias{merge_machine_readout_expr}
\title{Merge readout expressions during oapply_machine composition}
\usage{
merge_machine_readout_expr(d, machines, all_snames, offsets)
}
\description{
Merge readout expressions during oapply_machine composition
}
\keyword{internal}
