% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{merge_rs_expressions}
\alias{merge_rs_expressions}
\title{Merge symbolic expressions during oapply_dynam composition}
\usage{
merge_rs_expressions(
  sharers,
  all_snames,
  composite_names,
  composite_map,
  inverse_map,
  n_composite,
  nboxes,
  box_states
)
}
\description{
Merge symbolic expressions during oapply_dynam composition
}
\keyword{internal}
