% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{oapply}
\alias{oapply}
\title{Compose open Petri nets over a UWD}
\usage{
oapply(w, components)
}
\arguments{
\item{w}{A UWD ACSet}

\item{components}{List of Open Petri nets, one per box}
}
\value{
An Open Petri net (the composed system)
}
\description{
The UWD specifies how boxes (= open PNs) connect via junctions.
Each box's ports connect to junctions; shared junctions mean species
are identified (glued together).
}
