% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{oapply_machine}
\alias{oapply_machine}
\title{Compose Machines via a directed wiring diagram}
\usage{
oapply_machine(d, machines)
}
\arguments{
\item{d}{A directed wiring diagram (from \code{dwd()})}

\item{machines}{A list of Machine objects, one per box}
}
\value{
A composite Machine
}
\description{
Each box of the directed wiring diagram is filled by a Machine.
Outputs of upstream machines are wired to inputs of downstream ones.
}
