% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{Open}
\alias{Open}
\title{Open Petri net: a Petri net with exposed "legs" (species accessible
from outside for composition).}
\usage{
Open(pn, legs = NULL)
}
\arguments{
\item{pn}{A Petri net ACSet}

\item{legs}{List of integer vectors mapping outer junctions to species IDs}
}
\description{
Open Petri net: a Petri net with exposed "legs" (species accessible
from outside for composition).
}
