% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_petri_net}
\alias{open_petri_net}
\title{Open a Petri net at specified species, returning a StructuredCospan}
\usage{
open_petri_net(pn, ...)
}
\arguments{
\item{pn}{A Petri net ACSet}

\item{...}{Integer vectors, each specifying a leg (species indices)}
}
\value{
A catlab::StructuredCospan with interface_ob = "S"
}
\description{
Convenience wrapper around catlab::open_acset with interface_ob = "S".
}
\examples{
sir <- labelled_petri_net(
  c("S", "I", "R"),
  inf = c("S", "I") \%=>\% c("I", "I"),
  rec = "I" \%=>\% "R"
)
sc <- open_petri_net(sir, c(1L, 2L), c(2L, 3L))
}
