% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{petri_to_continuous}
\alias{petri_to_continuous}
\title{Convert a Petri net to a ContinuousResourceSharer (ODE)}
\usage{
petri_to_continuous(pn)
}
\arguments{
\item{pn}{A Petri net ACSet}
}
\value{
A ResourceSharer with system_type "continuous"
}
\description{
Convert a Petri net to a ContinuousResourceSharer (ODE)
}
