% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{petri_to_discrete}
\alias{petri_to_discrete}
\title{Convert a Petri net to a DiscreteResourceSharer (stochastic)}
\usage{
petri_to_discrete(pn)
}
\arguments{
\item{pn}{A Petri net ACSet}
}
\value{
A ResourceSharer with system_type "discrete"
}
\description{
Convert a Petri net to a DiscreteResourceSharer (stochastic)
}
