% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{petri_to_sf}
\alias{petri_to_sf}
\title{Convert a Petri net to a stock-flow model}
\usage{
petri_to_sf(pn)
}
\arguments{
\item{pn}{A LabelledPetriNet}
}
\value{
A StockFlowModel
}
\description{
Creates a basic stock-flow model from a Petri net. Flow rate functions
default to mass-action kinetics.
}
