% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\name{PetriNet}
\alias{PetriNet}
\title{PetriNet ACSet type (unlabelled)}
\usage{
PetriNet(...)
}
\arguments{
\item{...}{Arguments passed to the ACSet constructor}
}
\description{
PetriNet ACSet type (unlabelled)
}
\examples{
pn <- PetriNet()
s1 <- acsets::add_part(pn, "S")
t1 <- acsets::add_part(pn, "T")
acsets::add_part(pn, "I", is = s1, it = t1)
acsets::add_part(pn, "O", os = s1, ot = t1)
acsets::nparts(pn, "S") # 1
}
