% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot model fit against data}
\usage{
plot_fit(
  samples,
  data,
  obs_var,
  state_var = NULL,
  burnin = 0,
  quantiles = c(0.025, 0.25, 0.75, 0.975)
)
}
\arguments{
\item{samples}{A monty_samples object with saved trajectories}

\item{data}{The observed data (data.frame with time and data columns)}

\item{obs_var}{Name of the observed variable column in data}

\item{state_var}{Name of the state variable to plot from trajectories}

\item{burnin}{Number of initial samples to discard}

\item{quantiles}{Quantile levels for credible intervals (default:
50\% and 95\%)}
}
\value{
A ggplot object
}
\description{
Plots observed data with model predictions from MCMC posterior.
Requires that \code{save_trajectories = TRUE} was used in \code{\link[=fit_mcmc]{fit_mcmc()}}.
}
