% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_petri}
\alias{plot_petri}
\title{Plot a Petri net as a bipartite graph via DiagrammeR}
\usage{
plot_petri(pn)
}
\arguments{
\item{pn}{A Petri net ACSet, TypedPetriNet, or Open Petri net}
}
\value{
A DiagrammeR \code{htmlwidget}
}
\description{
Species are shown as blue circles, transitions as orange boxes.
Accepts LabelledPetriNet, TypedPetriNet, or Open Petri nets.
}
