% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior density estimates}
\usage{
plot_posterior(samples, pars = NULL, burnin = 0, true_values = NULL)
}
\arguments{
\item{samples}{A monty_samples object from \code{\link[=fit_mcmc]{fit_mcmc()}}}

\item{pars}{Optional character vector of parameter names to plot}

\item{burnin}{Number of initial samples to discard}

\item{true_values}{Optional named list of true parameter values
(shown as vertical lines)}
}
\value{
A ggplot object
}
\description{
Plot posterior density estimates
}
