% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{plot_stock_flow}
\alias{plot_stock_flow}
\title{Plot a stock-flow diagram using DiagrammeR}
\usage{
plot_stock_flow(sf)
}
\arguments{
\item{sf}{A StockFlowModel}
}
\value{
A DiagrammeR graph
}
\description{
Plot a stock-flow diagram using DiagrammeR
}
