% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot MCMC trace plots}
\usage{
plot_traces(samples, pars = NULL, burnin = 0)
}
\arguments{
\item{samples}{A monty_samples object from \code{\link[=fit_mcmc]{fit_mcmc()}}}

\item{pars}{Optional character vector of parameter names to plot}

\item{burnin}{Number of initial samples to shade as burn-in}
}
\value{
A ggplot object
}
\description{
Plot MCMC trace plots
}
