% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_uwd}
\alias{plot_uwd}
\title{Plot an undirected wiring diagram via DiagrammeR}
\usage{
plot_uwd(w)
}
\arguments{
\item{w}{A UWD ACSet}
}
\value{
A DiagrammeR \code{htmlwidget}
}
\description{
Boxes are yellow rectangles, junctions are grey circles,
outer ports are small black diamonds.
}
\examples{
w <- catlab::uwd(
  outer = c("s", "i", "r"),
  infection = c("s", "i"),
  recovery = c("i", "r")
)
if (requireNamespace("DiagrammeR", quietly = TRUE)) {
  plot_uwd(w)
}
}
