% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{prior_exp}
\alias{prior_exp}
\title{Exponential prior}
\usage{
prior_exp(mean = 1)
}
\arguments{
\item{mean}{Mean of the exponential distribution (default 1)}
}
\value{
A prior specification for \code{\link[=build_prior]{build_prior()}}
}
\description{
Exponential prior
}
