% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{prior_lognormal}
\alias{prior_lognormal}
\title{Log-normal prior}
\usage{
prior_lognormal(meanlog = 0, sdlog = 1)
}
\arguments{
\item{meanlog}{Mean of the log (default 0)}

\item{sdlog}{Standard deviation of the log (default 1)}
}
\value{
A prior specification for \code{\link[=build_prior]{build_prior()}}
}
\description{
Log-normal prior
}
