% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{prior_normal}
\alias{prior_normal}
\title{Normal prior}
\usage{
prior_normal(mean = 0, sd = 1)
}
\arguments{
\item{mean}{Mean (default 0)}

\item{sd}{Standard deviation (default 1)}
}
\value{
A prior specification for \code{\link[=build_prior]{build_prior()}}
}
\description{
Normal prior
}
