% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{prior_uniform}
\alias{prior_uniform}
\title{Uniform prior}
\usage{
prior_uniform(min = 0, max = 1)
}
\arguments{
\item{min}{Lower bound (default 0)}

\item{max}{Upper bound (default 1)}
}
\value{
A prior specification for \code{\link[=build_prior]{build_prior()}}
}
\description{
Uniform prior
}
