% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\name{ReactionNet}
\alias{ReactionNet}
\title{ReactionNet ACSet type (labelled with rates and concentrations)}
\usage{
ReactionNet(...)
}
\arguments{
\item{...}{Arguments passed to the ACSet constructor}
}
\description{
ReactionNet ACSet type (labelled with rates and concentrations)
}
