% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\docType{data}
\name{SchLabelledPetriNet}
\alias{SchLabelledPetriNet}
\title{Labelled Petri net schema}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchLabelledPetriNet
}
\description{
Labelled Petri net schema
}
\keyword{datasets}
