% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\docType{data}
\name{SchPetriNet}
\alias{SchPetriNet}
\title{Petri net schema (unlabelled)}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchPetriNet
}
\description{
Petri net schema (unlabelled)
}
\keyword{datasets}
