% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\docType{data}
\name{SchReactionNet}
\alias{SchReactionNet}
\title{Reaction net schema (labelled + rates + concentrations)}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchReactionNet
}
\description{
Reaction net schema (labelled + rates + concentrations)
}
\keyword{datasets}
