% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\docType{data}
\name{SchStockFlow}
\alias{SchStockFlow}
\title{Full Stock-and-Flow schema}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchStockFlow
}
\description{
Includes stocks, flows, inflows, outflows, auxiliary variables,
sum variables, parameters, and all link types.
}
\keyword{datasets}
