% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\docType{data}
\name{SchStockFlow0}
\alias{SchStockFlow0}
\title{Stock-and-Flow base schema (stocks + sum variables only)}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchStockFlow0
}
\description{
Used as the "foot" type for open stock-flow composition.
}
\keyword{datasets}
