% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_build_stratified}
\alias{sf_build_stratified}
\title{Build a stratified StockFlowModel from raw components}
\usage{
sf_build_stratified(stocks, flow_specs, params, sums)
}
\description{
Build a stratified StockFlowModel from raw components
}
\keyword{internal}
