% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_fnames}
\alias{sf_fnames}
\title{Get flow names}
\usage{
sf_fnames(sf)
}
\arguments{
\item{sf}{A StockFlowModel or StockFlow ACSet}
}
\value{
Character vector of flow names
}
\description{
Get flow names
}
