% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_infectious_type}
\alias{sf_infectious_type}
\title{SF type template for simple infectious disease models}
\usage{
sf_infectious_type()
}
\value{
A StockFlowModel usable as a type template
}
\description{
One stock type (pop), with flow types for disease transitions and
strata transitions. Analogous to \code{infectious_ontology()}.
}
