% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_oapply}
\alias{sf_oapply}
\title{Compose open stock-flow models via a UWD}
\usage{
sf_oapply(d, open_sfs)
}
\arguments{
\item{d}{A UWD ACSet (from catlab)}

\item{open_sfs}{A list of open stock-flow models}
}
\value{
A StockFlowModel (the composed model)
}
\description{
Compose open stock-flow models via a UWD
}
