% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_pnames}
\alias{sf_pnames}
\title{Get parameter names}
\usage{
sf_pnames(sf)
}
\arguments{
\item{sf}{A StockFlowModel or StockFlow ACSet}
}
\value{
Character vector of parameter names
}
\description{
Get parameter names
}
