% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_snames}
\alias{sf_snames}
\title{Get stock names}
\usage{
sf_snames(sf)
}
\arguments{
\item{sf}{A StockFlowModel or StockFlow ACSet}
}
\value{
Character vector of stock names
}
\description{
Get stock names
}
