% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_svnames}
\alias{sf_svnames}
\title{Get sum variable names}
\usage{
sf_svnames(sf)
}
\arguments{
\item{sf}{A StockFlowModel or StockFlow ACSet}
}
\value{
Character vector of sum variable names
}
\description{
Get sum variable names
}
