% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_to_odin_array_system}
\alias{sf_to_odin_array_system}
\title{Generate an array-based odin2 system from a stratified model}
\usage{
sf_to_odin_array_system(sf, type = "ode", initial = NULL, compare = NULL)
}
\arguments{
\item{sf}{A stratified StockFlowModel (from \code{\link[=sf_stratify]{sf_stratify()}})}

\item{type}{One of "ode", "stochastic", "discrete"}

\item{initial}{Optional named list of initial values}

\item{compare}{Optional named list of observation specifications}
}
\value{
A function that returns an odin2 generator
}
\description{
Generate an array-based odin2 system from a stratified model
}
