% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_to_petri}
\alias{sf_to_petri}
\title{Convert a stock-flow model to a labelled Petri net}
\usage{
sf_to_petri(sf)
}
\arguments{
\item{sf}{A StockFlowModel}
}
\value{
A LabelledPetriNet ACSet
}
\description{
Stocks become species, flows become transitions.
Inflows/outflows become input/output arcs.
}
