% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_to_resource_sharer}
\alias{sf_to_resource_sharer}
\title{Convert a stock-flow model to a continuous ResourceSharer}
\usage{
sf_to_resource_sharer(sf)
}
\arguments{
\item{sf}{A StockFlowModel}
}
\value{
A ResourceSharer with system_type = "continuous"
}
\description{
Convert a stock-flow model to a continuous ResourceSharer
}
