% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_typed}
\alias{sf_typed}
\title{Create a typed stock-flow model}
\usage{
sf_typed(
  sf,
  type_sf,
  stock_types = NULL,
  flow_types = NULL,
  param_types = NULL
)
}
\arguments{
\item{sf}{A StockFlowModel}

\item{type_sf}{A StockFlowModel acting as type template}

\item{stock_types}{Named character: sf stock name -> type stock name}

\item{flow_types}{Named character: sf flow name -> type flow name}

\item{param_types}{Named character: sf param name -> type param name}
}
\value{
A TypedStockFlow
}
\description{
Maps stocks, flows, and parameters of a stock-flow model to a type template.
}
