% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_vectorfield}
\alias{sf_vectorfield}
\title{Generate a vectorfield function from a stock-flow model}
\usage{
sf_vectorfield(sf)
}
\arguments{
\item{sf}{A StockFlowModel}
}
\value{
A function with signature (t, state, parms) suitable for deSolve
}
\description{
Generate a vectorfield function from a stock-flow model
}
\examples{
sir_sf <- stock_and_flow(
  stocks = c("S", "I", "R"),
  flows = list(
    inf = flow(from = "S", to = "I", rate = quote(beta * S * I / N)),
    rec = flow(from = "I", to = "R", rate = quote(gamma * I))
  ),
  params = c("beta", "gamma"),
  sums = list(N = c("S", "I", "R"))
)
vf <- sf_vectorfield(sir_sf)
state <- c(S = 990, I = 10, R = 0)
parms <- list(beta = 0.3, gamma = 0.1)
dstate <- vf(0, state, parms)
dstate[[1]] # derivatives: dS/dt, dI/dt, dR/dt
}
