% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_vnames}
\alias{sf_vnames}
\title{Get variable names}
\usage{
sf_vnames(sf)
}
\arguments{
\item{sf}{A StockFlowModel or StockFlow ACSet}
}
\value{
Character vector of variable names
}
\description{
Get variable names
}
