% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{simulate_machine}
\alias{simulate_machine}
\title{Simulate a Machine with external inputs}
\usage{
simulate_machine(
  m,
  initial,
  times,
  input_fn = function(t) numeric(m@ninputs),
  params = list()
)
}
\arguments{
\item{m}{A continuous Machine}

\item{initial}{Named numeric vector of initial state}

\item{times}{Numeric vector of output times}

\item{input_fn}{Function (t) -> numeric vector of inputs}

\item{params}{Parameter list}
}
\value{
A data.frame
}
\description{
Simulate a Machine with external inputs
}
