% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{simulate_rs}
\alias{simulate_rs}
\title{Simulate a continuous ResourceSharer using deSolve}
\usage{
simulate_rs(rs, initial, times, params = list())
}
\arguments{
\item{rs}{A continuous ResourceSharer}

\item{initial}{Named numeric vector of initial conditions}

\item{times}{Numeric vector of output times}

\item{params}{Parameter list (passed to dynamics as p)}
}
\value{
A data.frame with time and state columns
}
\description{
Simulate a continuous ResourceSharer using deSolve
}
