% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{simulate_rs_discrete}
\alias{simulate_rs_discrete}
\title{Simulate a discrete ResourceSharer}
\usage{
simulate_rs_discrete(rs, initial, nsteps, params = list())
}
\arguments{
\item{rs}{A discrete ResourceSharer}

\item{initial}{Named numeric vector of initial conditions}

\item{nsteps}{Number of time steps}

\item{params}{Parameter list}
}
\value{
A data.frame with time and state columns
}
\description{
Simulate a discrete ResourceSharer
}
