% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{simulate_sf}
\alias{simulate_sf}
\title{Simulate a stock-flow model using deSolve}
\usage{
simulate_sf(sf, initial, times, params)
}
\arguments{
\item{sf}{A StockFlowModel}

\item{initial}{Named numeric vector of initial conditions}

\item{times}{Numeric vector of output times}

\item{params}{Named list of parameter values}
}
\value{
A data.frame
}
\description{
Simulate a stock-flow model using deSolve
}
