% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\name{species_names}
\alias{species_names}
\title{Get species names from a Petri net}
\usage{
species_names(pn)
}
\arguments{
\item{pn}{A Petri net ACSet (or TypedPetriNet)}
}
\value{
Character vector of species names
}
\description{
Get species names from a Petri net
}
