% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{spontaneous_petri}
\alias{spontaneous_petri}
\title{Spontaneous transition: A -> B}
\usage{
spontaneous_petri(input, output, tname = NULL)
}
\arguments{
\item{input}{Name of the input species}

\item{output}{Name of the output species}

\item{tname}{Optional transition name}
}
\value{
An Open Petri net
}
\description{
Spontaneous transition: A -> B
}
