% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{StockFlow0}
\alias{StockFlow0}
\title{Create StockFlow0 ACSet type (for composition feet)}
\usage{
StockFlow0(...)
}
\arguments{
\item{...}{Arguments passed to the ACSet constructor}
}
\description{
Create StockFlow0 ACSet type (for composition feet)
}
