% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{StockFlowModel}
\alias{StockFlowModel}
\title{A Stock-and-Flow model}
\usage{
StockFlowModel(acset = NULL, flow_fns = list(), var_exprs = list())
}
\arguments{
\item{acset}{The underlying StockFlow ACSet}

\item{flow_fns}{Named list of internal flow rate functions:
name -> function(u, sv, p, t)}

\item{var_exprs}{Named list of variable expression strings (for display)}
}
\description{
Wraps a StockFlow ACSet with additional metadata: flow rate functions
(expressed as R expressions or functions) and operator information.
}
