% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{StockFlowType}
\alias{StockFlowType}
\title{Create StockFlow ACSet type}
\usage{
StockFlowType(...)
}
\arguments{
\item{...}{Arguments passed to the ACSet constructor}
}
\description{
Create StockFlow ACSet type
}
