% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{User-friendly stratification wrapper}
\usage{
stratify(base, strata)
}
\arguments{
\item{base}{A \code{TypedPetriNet} (the base model, e.g. SIR)}

\item{strata}{A \code{TypedPetriNet} (the stratification model, e.g. age groups)}
}
\value{
A \code{TypedPetriNet} (the stratified model)
}
\description{
Computes the typed product of two TypedPetriNets.
Equivalent to \code{typed_product(base, strata)}.
}
