% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{subst_exprs}
\alias{subst_exprs}
\title{Substitute symbolic sub-expressions into an expression}
\usage{
subst_exprs(expr, mapping)
}
\description{
Substitute symbolic sub-expressions into an expression
}
\keyword{internal}
