% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{sum_exprs}
\alias{sum_exprs}
\title{Sum a list of expressions into a single expression}
\usage{
sum_exprs(exprs)
}
\arguments{
\item{exprs}{List of language objects to sum}
}
\value{
A single expression representing the sum
}
\description{
Sum a list of expressions into a single expression
}
\keyword{internal}
