% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{transition_matrices}
\alias{transition_matrices}
\title{Extract input/output stoichiometry matrices}
\usage{
transition_matrices(pn)
}
\arguments{
\item{pn}{A Petri net ACSet}
}
\value{
A list with \code{input} and \code{output} matrices (species × transitions)
}
\description{
Extract input/output stoichiometry matrices
}
\examples{
sir <- labelled_petri_net(
  c("S", "I", "R"),
  inf = c("S", "I") \%=>\% c("I", "I"),
  rec = "I" \%=>\% "R"
)
tm <- transition_matrices(sir)
tm$input         # species x transitions input matrix
tm$output        # species x transitions output matrix
tm$stoichiometry # net change: output - input
}
