% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\name{transition_names}
\alias{transition_names}
\title{Get transition names from a Petri net}
\usage{
transition_names(pn)
}
\arguments{
\item{pn}{A Petri net ACSet (or TypedPetriNet)}
}
\value{
Character vector of transition names
}
\description{
Get transition names from a Petri net
}
