% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{typed_petri}
\alias{typed_petri}
\title{Create a typed Petri net from type assignments}
\usage{
typed_petri(pn, type_system, species_types = NULL, transition_types = NULL)
}
\arguments{
\item{pn}{A LabelledPetriNet}

\item{type_system}{A LabelledPetriNet (the ontology)}

\item{species_types}{Named character vector mapping species names to
type system species names, e.g. \code{c(S = "Pop", I = "Pop")}}

\item{transition_types}{Named character vector mapping transition names
to type system transition names, e.g. \code{c(inf = "infect", rec = "disease")}}
}
\value{
A \code{TypedPetriNet}
}
\description{
Create a typed Petri net from type assignments
}
