% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{typed_product}
\alias{typed_product}
\title{Compute the typed product (stratification) of typed Petri nets}
\usage{
typed_product(tp1, tp2)
}
\arguments{
\item{tp1, tp2}{\code{TypedPetriNet} objects with the same type system}
}
\value{
A \code{TypedPetriNet} (the stratified model)
}
\description{
This is the core stratification operation. Given two typed Petri nets
sharing the same type system, the typed product creates a new Petri net
whose species are pairs of species with matching types, and whose
transitions are pairs of transitions with matching types.
}
