% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{apex}
\alias{apex}
\title{Extract the closed Petri net from an Open Petri net}
\usage{
apex(open_pn)
}
\arguments{
\item{open_pn}{An Open Petri net}
}
\value{
A Petri net ACSet
}
\description{
Extract the closed Petri net from an Open Petri net
}
\examples{
pn <- labelled_petri_net(c("S", "I"), inf = "S" \%=>\% "I")
open_pn <- Open(pn)
closed <- apex(open_pn)
species_names(closed) # c("S", "I")
}
