% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{as_open}
\alias{as_open}
\title{Convert a catlab StructuredCospan back to an Open Petri net}
\usage{
as_open(sc)
}
\arguments{
\item{sc}{A catlab::StructuredCospan}
}
\value{
An Open Petri net
}
\description{
Extracts the apex and recovers each leg as an integer vector of species
indices from the ACSetTransformation components.
}
\examples{
sir <- labelled_petri_net(
  c("S", "I", "R"),
  inf = c("S", "I") \%=>\% c("I", "I"),
  rec = "I" \%=>\% "R"
)
sc <- open_petri_net(sir, 1:3)
open_sir <- as_open(sc)
species_names(apex(open_sir))
}
