% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{as_structured_cospan}
\alias{as_structured_cospan}
\title{Convert an Open Petri net to a catlab StructuredCospan}
\usage{
as_structured_cospan(open_pn)
}
\arguments{
\item{open_pn}{An Open Petri net}
}
\value{
A catlab::StructuredCospan with interface_ob = "S"
}
\description{
Each integer-vector leg becomes an ACSetTransformation from a discrete
Petri net foot (with parts only in "S") to the apex.
}
\examples{
sir <- labelled_petri_net(
  c("S", "I", "R"),
  inf = c("S", "I") \%=>\% c("I", "I"),
  rec = "I" \%=>\% "R"
)
open_sir <- Open(sir)
sc <- as_structured_cospan(open_sir)
# Round-trip back to Open
open_again <- as_open(sc)
species_names(apex(open_again))
}
