% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{compose}
\alias{compose}
\title{Compose multiple Open Petri nets}
\usage{
compose(...)
}
\arguments{
\item{...}{Open Petri nets}
}
\value{
An Open Petri net
}
\description{
Convenience function for composing more than two Open Petri nets
by iterative pairwise composition.
}
