% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{compose_epi}
\alias{compose_epi}
\title{Compose an epidemiological model from a UWD and dictionary}
\usage{
compose_epi(w, dict, box_names)
}
\arguments{
\item{w}{A UWD (wiring diagram)}

\item{dict}{A named list of Open Petri nets, keyed by box names}

\item{box_names}{Character vector mapping box indices to dict keys}
}
\value{
An Open Petri net
}
\description{
Compose an epidemiological model from a UWD and dictionary
}
