% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{compose_open}
\alias{compose_open}
\title{Compose two Open Petri nets by matching species names}
\usage{
compose_open(a, b)
}
\arguments{
\item{a, b}{Open Petri nets}
}
\value{
An Open Petri net
}
\description{
Creates a UWD where species with the same name are identified (shared
junctions), then composes via \code{oapply}.
}
