% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{exposure_petri}
\alias{exposure_petri}
\title{Exposure/contact transition: A + B -> C + B (B catalyzes A -> C)}
\usage{
exposure_petri(susceptible, infectious, output, tname = NULL)
}
\arguments{
\item{susceptible}{Name of the susceptible species}

\item{infectious}{Name of the infectious species}

\item{output}{Name of the output species}

\item{tname}{Optional transition name}
}
\value{
An Open Petri net
}
\description{
Exposure/contact transition: A + B -> C + B (B catalyzes A -> C)
}
