% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\name{\%=>\%}
\alias{\%=>\%}
\title{Define a Petri net transition: inputs to outputs}
\usage{
inputs \%=>\% outputs
}
\arguments{
\item{inputs}{Character vector of input species names}

\item{outputs}{Character vector of output species names}
}
\value{
A list with \code{inputs} and \code{outputs}
}
\description{
Define a Petri net transition: inputs to outputs
}
