% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\name{labelled_petri_net}
\alias{labelled_petri_net}
\title{Build a LabelledPetriNet from species names and transition specs}
\usage{
labelled_petri_net(species, ...)
}
\arguments{
\item{species}{Character vector of species names}

\item{...}{Named transition specs: \code{name = inputs \%=>\% outputs}
where inputs/outputs are character vectors of species names}
}
\value{
A LabelledPetriNet ACSet
}
\description{
Build a LabelledPetriNet from species names and transition specs
}
