% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\name{LabelledPetriNet}
\alias{LabelledPetriNet}
\title{LabelledPetriNet ACSet type}
\usage{
LabelledPetriNet(...)
}
\arguments{
\item{...}{Arguments passed to the ACSet constructor}
}
\description{
LabelledPetriNet ACSet type
}
\examples{
pn <- LabelledPetriNet()
s1 <- acsets::add_part(pn, "S", sname = "S")
s2 <- acsets::add_part(pn, "S", sname = "I")
t1 <- acsets::add_part(pn, "T", tname = "inf")
acsets::add_part(pn, "I", is = s1, it = t1)
acsets::add_part(pn, "O", os = s2, ot = t1)
species_names(pn) # c("S", "I")
}
