% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{Open}
\alias{Open}
\title{Open Petri net: a Petri net with exposed "legs" (species accessible
from outside for composition).}
\usage{
Open(pn, legs = NULL)
}
\arguments{
\item{pn}{A Petri net ACSet}

\item{legs}{List of integer vectors mapping outer junctions to species IDs}
}
\description{
Open Petri net: a Petri net with exposed "legs" (species accessible
from outside for composition).
}
\examples{
sir <- labelled_petri_net(
  c("S", "I", "R"),
  inf = c("S", "I") \%=>\% c("I", "I"),
  rec = "I" \%=>\% "R"
)
# Expose all species as a single leg
open_sir <- Open(sir)
species_names(apex(open_sir)) # c("S", "I", "R")

# Expose specific species
open_si <- Open(sir, legs = list(c(1L, 2L)))
}
