% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynam.R
\name{petri_to_continuous}
\alias{petri_to_continuous}
\title{Convert a Petri net to a ContinuousResourceSharer (ODE)}
\usage{
petri_to_continuous(pn)
}
\arguments{
\item{pn}{A Petri net ACSet}
}
\value{
A ResourceSharer with system_type "continuous"
}
\description{
Convert a Petri net to a ContinuousResourceSharer (ODE)
}
\examples{
sir <- labelled_petri_net(
  c("S", "I", "R"),
  inf = c("S", "I") \%=>\% c("I", "I"),
  rec = "I" \%=>\% "R"
)
rs <- petri_to_continuous(sir)
rs@system_type   # "continuous"
rs@state_names   # c("S", "I", "R")
}
