% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\name{PetriNet}
\alias{PetriNet}
\title{PetriNet ACSet type (unlabelled)}
\usage{
PetriNet(...)
}
\arguments{
\item{...}{Arguments passed to the ACSet constructor}
}
\description{
PetriNet ACSet type (unlabelled)
}
