% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diff}
\alias{plot_diff}
\title{Plot absolute differences between two ODE solutions}
\usage{
plot_diff(sol1, sol2, vars = NULL, title = "Absolute Difference")
}
\arguments{
\item{sol1}{First solution (data.frame)}

\item{sol2}{Second solution (data.frame)}

\item{vars}{Character vector of variable names}

\item{title}{Optional plot title}
}
\value{
A ggplot2 object
}
\description{
Plot absolute differences between two ODE solutions
}
\examples{
t <- seq(0, 10, 0.1)
sol1 <- data.frame(time = t, S = 990 * exp(-0.3 * t), I = 10 * exp(0.1 * t))
sol2 <- data.frame(time = t, S = 989 * exp(-0.3 * t), I = 11 * exp(0.1 * t))
if (requireNamespace("ggplot2", quietly = TRUE)) {
  plot_diff(sol1, sol2)
}
}
