% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.R
\name{ReactionNet}
\alias{ReactionNet}
\title{ReactionNet ACSet type (labelled with rates and concentrations)}
\usage{
ReactionNet(...)
}
\arguments{
\item{...}{Arguments passed to the ACSet constructor}
}
\description{
ReactionNet ACSet type (labelled with rates and concentrations)
}
\examples{
rn <- ReactionNet()
s1 <- acsets::add_part(rn, "S", sname = "S", concentration = 990)
s2 <- acsets::add_part(rn, "S", sname = "I", concentration = 10)
t1 <- acsets::add_part(rn, "T", tname = "inf", rate = 0.001)
acsets::add_part(rn, "I", is = s1, it = t1)
acsets::add_part(rn, "O", os = s2, ot = t1)
}
