% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_transition_matrices}
\alias{sf_transition_matrices}
\title{Compute inflow/outflow transition matrices for a stock-flow model}
\usage{
sf_transition_matrices(sf)
}
\arguments{
\item{sf}{A StockFlowModel}
}
\value{
A list with \code{inflow} and \code{outflow} matrices
(rows = flows, cols = stocks)
}
\description{
Compute inflow/outflow transition matrices for a stock-flow model
}
