% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{sf_vectorfield}
\alias{sf_vectorfield}
\title{Generate a vectorfield function from a stock-flow model}
\usage{
sf_vectorfield(sf)
}
\arguments{
\item{sf}{A StockFlowModel}
}
\value{
A function with signature (t, state, parms) suitable for deSolve
}
\description{
Generate a vectorfield function from a stock-flow model
}
