% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{transition_matrices}
\alias{transition_matrices}
\title{Extract input/output stoichiometry matrices}
\usage{
transition_matrices(pn)
}
\arguments{
\item{pn}{A Petri net ACSet}
}
\value{
A list with \code{input} and \code{output} matrices (species × transitions)
}
\description{
Extract input/output stoichiometry matrices
}
